remove unused details code
