+++ composeApp/src/androidMain/kotlin/com/dnfapps/arrmatey/ui/components/UnifiedMediaInfoItems.kt
@@ -1,27 +1,29 @@
 package com.dnfapps.arrmatey.ui.components
 
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalLocale
 import com.dnfapps.arrmatey.arr.api.model.ArrMovie
 import com.dnfapps.arrmatey.arr.api.model.ArrSeries
 import com.dnfapps.arrmatey.arr.api.model.Arrtist
+import com.dnfapps.arrmatey.arr.api.model.ArtistMonitorType
 import com.dnfapps.arrmatey.arr.api.model.Audiobook
 import com.dnfapps.arrmatey.arr.api.model.Author
+import com.dnfapps.arrmatey.arr.api.model.AuthorMonitorType
+import com.dnfapps.arrmatey.arr.api.model.MonitorNewItems
 import com.dnfapps.arrmatey.arr.api.model.QualityProfile
 import com.dnfapps.arrmatey.arr.api.model.Tag
+import com.dnfapps.arrmatey.compose.utils.bytesAsFileSizeString
 import com.dnfapps.arrmatey.compose.utils.formatWithCommas
+import com.dnfapps.arrmatey.entensions.Bullet
 import com.dnfapps.arrmatey.model.InfoItem
 import com.dnfapps.arrmatey.model.UnifiedMediaDetailsUiState
 import com.dnfapps.arrmatey.model.toInfoList
 import com.dnfapps.arrmatey.seerr.api.model.MovieDetails
 import com.dnfapps.arrmatey.seerr.api.model.PersonDetails
 import com.dnfapps.arrmatey.shared.MR
-import com.dnfapps.arrmatey.ui.screens.artistInfo
-import com.dnfapps.arrmatey.ui.screens.audiobookInfo
-import com.dnfapps.arrmatey.ui.screens.authorInfo
-import com.dnfapps.arrmatey.ui.screens.movieInfo
-import com.dnfapps.arrmatey.ui.screens.seriesInfo
 import com.dnfapps.arrmatey.utils.format
-import dev.icerock.moko.resources.compose.stringResource as mokoString
+import com.dnfapps.arrmatey.utils.mokoString
+import kotlin.time.ExperimentalTime
 
 @Composable
 fun buildUnifiedInfoItems(
@@ -67,5 +69,157 @@ fun buildUnifiedInfoItems(
         if (studiosText.isNotEmpty()) {
             add(InfoItem(mokoString(MR.strings.studios), studiosText))
         }
+    }
+}
+
+@Composable
+fun seriesInfo(
+    series: ArrSeries,
+    qualityProfiles: List<QualityProfile>,
+    tags: List<Tag>
+): Map<String, String> {
+    val qualityProfile = qualityProfiles.firstOrNull { it.id == series.qualityProfileId }
+    val tagsLabel = series.formatTags(tags) ?: mokoString(MR.strings.none)
+
+    val unknown = mokoString(MR.strings.unknown)
+    val monitorLabel = if (series.monitorNewItems == MonitorNewItems.All) {
+        mokoString(MR.strings.monitored)
+    } else { mokoString(MR.strings.unmonitored) }
+
+    val seasonFolderLabel = if (series.seasonFolder) {
+        mokoString(MR.strings.yes)
+    } else { mokoString(MR.strings.no) }
+
+    val diskSize = series.fileSize.bytesAsFileSizeString()
+
+    return mapOf(
+        mokoString(MR.strings.status) to mokoString(series.status.resource),
+        mokoString(MR.strings.series_type) to series.seriesType.name,
+        mokoString(MR.strings.size_on_disk) to diskSize,
+        mokoString(MR.strings.root_folder) to (series.rootFolderPath ?: unknown),
+        mokoString(MR.strings.path) to (series.path ?: unknown),
+        mokoString(MR.strings.new_seasons) to monitorLabel,
+        mokoString(MR.strings.season_folders) to seasonFolderLabel,
+        mokoString(MR.strings.quality_profile) to (qualityProfile?.name ?: unknown),
+        mokoString(MR.strings.tags) to tagsLabel
+    )
+}
+
+@OptIn(ExperimentalTime::class)
+@Composable
+fun movieInfo(
+    movie: ArrMovie,
+    qualityProfiles: List<QualityProfile>,
+    tags: List<Tag>
+): Map<String, String> {
+    val qualityProfile = qualityProfiles.firstOrNull { it.id == movie.qualityProfileId }
+    val tagsLabel = movie.formatTags(tags) ?: mokoString(MR.strings.none)
+
+    val unknown = mokoString(MR.strings.unknown)
+
+    val rootFolderPathValue = movie.rootFolderPath.takeUnless { it.isBlank() }
+        ?: mokoString(MR.strings.unknown)
+
+    return buildMap {
+        put(mokoString(MR.strings.status), mokoString(movie.status.resource))
+        put(mokoString(MR.strings.minimum_availability), movie.minimumAvailability.name)
+        put(mokoString(MR.strings.root_folder), rootFolderPathValue)
+        put(mokoString(MR.strings.path), (movie.path ?: unknown))
+        movie.inCinemas?.format("MMM d, yyyy")?.let {
+            put(mokoString(MR.strings.in_cinemas), it)
+        }
+        movie.physicalRelease?.format("MMM d, yyyy")?.let {
+            put(mokoString(MR.strings.physical_release), it)
+        }
+        movie.digitalRelease?.format("MMM d, yyyy")?.let {
+            put(mokoString(MR.strings.digital_release), it)
+        }
+        put(mokoString(MR.strings.quality_profile), (qualityProfile?.name ?: unknown))
+        put(mokoString(MR.strings.tags), tagsLabel)
+    }
+
+}
+
+@Composable
+fun artistInfo(
+    artist: Arrtist,
+    qualityProfiles: List<QualityProfile>,
+    tags: List<Tag>
+): Map<String, String> {
+    val qualityProfile = qualityProfiles.firstOrNull { it.id == artist.qualityProfileId }
+    val tagsLabel = artist.formatTags(tags) ?: mokoString(MR.strings.none)
+
+    val unknown = mokoString(MR.strings.unknown)
+    val monitorLabel = if (artist.monitorNewItems == ArtistMonitorType.All) {
+        mokoString(MR.strings.monitored)
+    } else { mokoString(MR.strings.unmonitored) }
+
+    val rootFolderPathValue = artist.rootFolderPath?.takeUnless { it.isBlank() }
+        ?: mokoString(MR.strings.unknown)
+
+    val diskSize = artist.fileSize.bytesAsFileSizeString()
+
+    return buildMap {
+        put(mokoString(MR.strings.status), mokoString(artist.status.resource))
+        put(mokoString(MR.strings.size_on_disk), diskSize)
+        put(mokoString(MR.strings.root_folder), rootFolderPathValue)
+        put(mokoString(MR.strings.path), (artist.path ?: unknown))
+        put(mokoString(MR.strings.new_albums), monitorLabel)
+        put(mokoString(MR.strings.quality_profile), (qualityProfile?.name ?: unknown))
+        put(mokoString(MR.strings.tags), tagsLabel)
+    }
+}
+
+@Composable
+fun authorInfo(
+    author: Author,
+    qualityProfiles: List<QualityProfile>,
+    tags: List<Tag>
+): Map<String, String> {
+    val qualityProfile = qualityProfiles.firstOrNull { it.id == author.qualityProfileId }
+    val tagsLabel = author.formatTags(tags) ?: mokoString(MR.strings.none)
+
+    val unknown = mokoString(MR.strings.unknown)
+    val monitorLabel = if (author.monitorNewItems == AuthorMonitorType.All) {
+        mokoString(MR.strings.monitored)
+    } else { mokoString(MR.strings.unmonitored) }
+
+    val rootFolderPathValue = author.rootFolderPath?.takeUnless { it.isBlank() }
+        ?: mokoString(MR.strings.unknown)
+
+    val diskSize = author.fileSize.bytesAsFileSizeString()
+
+    return buildMap {
+        put(mokoString(MR.strings.status), mokoString(author.status.resource))
+        put(mokoString(MR.strings.size_on_disk), diskSize)
+        put(mokoString(MR.strings.root_folder), rootFolderPathValue)
+        put(mokoString(MR.strings.path), (author.path ?: unknown))
+        put(mokoString(MR.strings.new_books), monitorLabel)
+        put(mokoString(MR.strings.quality_profile), (qualityProfile?.name ?: unknown))
+        put(mokoString(MR.strings.tags), tagsLabel)
+    }
+}
+
+@Composable
+fun audiobookInfo(
+    audiobook: Audiobook
+): Map<String, String> {
+    val unknown = mokoString(MR.strings.unknown)
+
+    val diskSize = audiobook.fileSize.bytesAsFileSizeString()
+
+    val authorString = audiobook.authors.takeUnless { it.isEmpty() }?.joinToString(Bullet) ?: unknown
+    val narratorsString = audiobook.narrators.takeUnless { it.isEmpty() }?.joinToString(Bullet) ?: unknown
+
+    return buildMap {
+        put(mokoString(MR.strings.audiobook_info_authors), authorString)
+        put(mokoString(MR.strings.audiobook_info_narrators), narratorsString)
+        put(mokoString(MR.strings.publisher), (audiobook.publisher ?: unknown))
+        audiobook.language?.let { language ->
+            put(mokoString(MR.strings.language),
+                language.replaceFirstChar { if (it.isLowerCase()) it.titlecase(LocalLocale.current.platformLocale) else it.toString() })
+        }
+        put(mokoString(MR.strings.size_on_disk), diskSize)
+        put(mokoString(MR.strings.path), (audiobook.path ?: unknown))
     }
 }
